fix: edit mistakes after code-review
