+++ src/const.js
@@ -1,0 +1,43 @@
+const TYPES = ['Taxi', 'Bus', 'Train', 'Ship', 'Transport', 'Drive', 'Flight', 'Check-in', 'Sightseeing', 'Restaurant'];
+const OFFERS = [
+  {
+    name: 'Add luggage',
+    price: 30,
+    shortName: 'lugagge',
+  },
+  {
+    name: 'Switch to comfort class',
+    price: 100,
+    shortName: 'comfort',
+  },
+  {
+    name: 'Add meal',
+    price: 15,
+    shortName: 'meal',
+  },
+  {
+    name: 'Choose seats',
+    price: 5,
+    shortName: 'seats',
+  },
+  {
+    name: 'Travel by train',
+    price: 40,
+    shortName: 'train',
+  },
+  {
+    name: 'Upgrade to a business class',
+    price: 120,
+    shortName: 'business',
+  },
+  {
+    name: 'Choose the radio station',
+    price: 60,
+    shortName: 'radio',
+  },
+];
+
+export {
+  TYPES,
+  OFFERS
+}
